+++ packages/theme_tailor/example/lib/diagnosticable_barrel_import.dart
@@ -1,10 +1,18 @@
+// ignore_for_file: annotate_overrides
+
 /// Import flutter foundation to generate debugFillProperties
 import 'package:example/diagnosticable_lib.dart';
 
 part 'diagnosticable_barrel_import.tailor.dart';
 
-@tailor
-class _$MyTheme {
-  static List<Color> background = [Colors.white, Colors.black];
-  static List<TextStyle> textStyle = const [TextStyle(), TextStyle()];
+@TailorMixin()
+class MyTheme extends ThemeExtension<MyTheme>
+    with DiagnosticableTreeMixin, _$MyThemeTailorMixin {
+  MyTheme({
+    required this.background,
+    required this.textStyle,
+  });
+
+  final Color background;
+  final TextStyle textStyle;
 }