+++ packages/theme_tailor/example/lib/empty_theme.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
-
-part 'empty_theme.tailor.dart';
-
-@Tailor(themes: [])
-class $_EmptyTheme {}