+++ packages/theme_tailor/example/lib/themes_name_collision.dart
@@ -1,15 +1,24 @@
+// ignore_for_file: annotate_overrides
+
 import 'package:flutter/material.dart';
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
 
 part 'themes_name_collision.tailor.dart';
 
-@tailor
-class $_SimpleTheme {
+@tailorMixin
+class SimpleTheme extends ThemeExtension<SimpleTheme>
+    with _$SimpleThemeTailorMixin {
   @themeExtension
-  static List<AnotherTheme> anotherTheme = AnotherTheme.tailorThemes;
+  static List<AnotherTheme> anotherTheme = [
+    const AnotherTheme(Colors.black),
+    const AnotherTheme(Colors.white),
+  ];
 }
 
-@tailorComponent
-class $_AnotherTheme {
-  static List<Color> themes = [Colors.amber, Colors.blueGrey.shade800];
+@tailorMixinComponent
+class AnotherTheme extends ThemeExtension<AnotherTheme>
+    with _$AnotherThemeTailorMixin {
+  const AnotherTheme(this.foreground);
+
+  final Color foreground;
 }