+++ packages/theme_tailor/lib/src/util/extension/element_annotation_extension.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use
+
 import 'package:analyzer/dart/element/element.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:source_gen_test/source_gen_test.dart';