+++ packages/theme_tailor/test/code_gen/inputs/field_keyword_generation_test_input.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use
+
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
 import 'package:source_gen_test/annotations.dart';