+++ packages/theme_tailor_annotation/lib/src/theme_tailor.g.dart
@@ -1,5 +1,7 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
+// ignore_for_file: deprecated_member_use_from_same_package
+
 part of 'theme_tailor.dart';
 
 // **************************************************************************
@@ -11,6 +13,8 @@ Tailor _$TailorFromJson(Map json) => Tailor(
           (json['themes'] as List<dynamic>?)?.map((e) => e as String).toList(),
       themeGetter:
           $enumDecodeNullable(_$ThemeGetterEnumMap, json['theme_getter']),
+      themeClassName: json['theme_class_name'] as String?,
+      themeDataClassName: json['theme_data_class_name'] as String?,
       requireStaticConst: json['require_static_const'] as bool?,
       generateStaticGetters: json['generate_static_getters'] as bool?,
     );