+++ packages/theme_tailor_annotation/lib/src/theme_tailor_component.dart
@@ -2,11 +2,13 @@ import 'package:meta/meta_meta.dart';
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
 
 /// Default instance of [TailorComponent] annotation
+@Deprecated('Use TailorMixinComponent instead')
 const tailorComponent = TailorComponent();
 
 /// An annotation used to specify that generated ThemeExtension class can be
 /// used as a part of other ThemeExtension class and does not need additional
 /// extensions on BuildContext nor ThemeData for the ease of access
+@Deprecated('Use TailorMixinComponent instead')
 @Target({TargetKind.classType})
 class TailorComponent extends Tailor {
   const TailorComponent({