+++ packages/theme_tailor/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.1.0
+- Add `themeClassName` and `themeDataClassName` properties to `@TailorMixin`. This addition provides the flexibility for themeGetters to generate a custom theme getter for the extension.
+- Add migration example from `@Tailor` / `@TailorComponent` to `@TailorMixin` / `@TailorMixinComponent`.
+- Bump dependencies versions
+
 # 2.0.2
 - Fix dependencies (Remove pubdev unknown platforms error)